requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
pandas==2.1.3
openpyxl==3.1.2
paddleocr==2.7.0
paddlepaddle==2.6.2
numpy==1.24.3
python-multipart==0.0.6
requests==2.31.0
pyyaml==6.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pillow==10.1.0
python-dotenv==1.0.0
